+++ CHANGELOG.md
@@ -11,6 +11,14 @@
 - changed/fixed - resume takes a map `(resume {:url url})`
 - fixed - `(restore)` is an alias of `(resotore {:url default-storage-url})`
 
+# 0.7.1 (2025-08-29)
+
+- FIXED: use of default storage.
+  (start) or (start {:scheme s}) starts on-memory.
+  (start {:url nil}) starts using `default-storage-url`.
+  (start {:url url}) create a new database store at url.
+  (restore) or (restart url) restart from memoried database.
+
 # 0.7.0 (2025-08-28)
 
 - resumed to MAJOR.MINOR.PATCH.